+++ app/helpers/application_helper.rb
@@ -40,26 +40,6 @@ module ApplicationHelper
     image_tag gravatar_url, alt: user.name, class: "gravatar"
   end
 
-  def sort_keys
-    {
-      t("search.sort_by_time") => "time",
-      t("search.sort_by_view") => "view",
-      t("search.sort_by_like") => "like",
-      t("search.sort_by_vote") => "vote"
-    }
-  end
-
-  def category_keys
-    if params.has_key?(:categories)
-      keys = params[:categories].values.reject { |c| c.empty? }.map(&:to_i)
-      return keys if keys.count > 0
-    end
-  end
-
-  def search_keyword
-    params[:q][:name_or_description_cont]
-  end
-
   def active_steps_form params, step
     if params.nil?
       case step
